Improve scripts

+++ deploy.bash
@@ -1,17 +1,26 @@
 #!/bin/bash
 
+# Build and deploy a new version to the remote device
+
 DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" >/dev/null 2>&1 && pwd )"
 re_ip='^[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}$'
 re_ip_user='^.+@[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}$'
 IP_LIST=""
 MVN_SETTINGS_FILE=""
+RESTART=0
 
 while test $# -gt 0
 do
     case "$1" in
         -h)
             shift
-            echo "Usage: ./deploy.bash [-H IP | -H USER@IP]..."
+            echo "Usage: ./deploy.bash [-H IP | -H USER@IP]... [-s SETTINGS_FILE] [-v VERSION]"
+            echo "      -h          Show this help message"
+            echo "      -H IP       Specify a host with username pi"
+            echo "      -H USER@IP  Specify a host with a different username"
+            echo "      -R          Restart the remote host instead of just restarting the service"
+            echo "      -s          Specify a maven settings file"
+            echo "      -v          Specify a specific version of animatedledstrip to build"
             exit
         ;;
         -H)
@@ -30,11 +39,23 @@ do
                 shift
             fi
         ;;
+        -R)
+            shift
+            RESTART=1
+        ;;
         -s)
             shift
             if test $# -gt 0
             then
                 MVN_SETTINGS_FILE=$1
+                shift
+            fi
+        ;;
+        -v)
+            shift
+            if test $# -gt 0
+            then
+                ALS_VERSION=$1
                 shift
             fi
         ;;
@@ -46,7 +67,12 @@ then
     MVN_SETTINGS_FILE="-gs ${MVN_SETTINGS_FILE}"
 fi
 
-mvn package ${MVN_SETTINGS_FILE}
+if [[ ${ALS_VERSION} != "" ]]
+then
+    ALS_VERSION="-Danimatedledstrip.version=${ALS_VERSION}"
+fi
+
+mvn package ${MVN_SETTINGS_FILE} ${ALS_VERSION}
 
 if [[ ${IP_LIST} == "" ]]
 then
@@ -59,7 +85,12 @@ else
         if [[ ${host} != "" ]]
         then
             scp ${DIR}/target/animatedledstrip-server-example-1.0.jar ${host}:/usr/local/leds/ledserver.jar
-            ssh ${host} -t "sudo reboot"
+            if [[ ${RESTART} == 1 ]]
+            then
+                ssh ${host} -t "sudo reboot"
+            else
+                ssh ${host} -t "sudo systemctl restart ledserver"
+            fi
         fi
     done
 fi
